requirements.txt
sentence-transformers==5.1.0
faiss-cpu==1.12.0
rank-bm25==0.2.2
scikit-learn==1.7.2
numpy==2.3.3
azure-ai-documentintelligence==1.0.2
requests==2.32.5
python-dotenv==1.1.1
streamlit